Rename kotlin package to org.bitcoindevkit, rust lib to bdkffi
