+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "uniffi_bdk"
+name = "bdk-ffi"
 version = "0.1.0"
 authors = ["Steve Myers <steve@notmandatory.org>", "Sudarsan Balaji <sudarsan.balaji@artfuldev.com>"]
 edition = "2018"
@@ -7,6 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
 crate-type = ["staticlib", "cdylib"]
+name = "bdkffi"
 
 [dependencies]
 bdk = { version = "0.13", features = ["all-keys", "use-esplora-ureq"] }
