+++ bindings/bdk-kotlin/demo/build.gradle
@@ -3,7 +3,7 @@ plugins {
     id 'application'
 }
 
-group = 'uniffi.bdk'
+group = 'org.bitcoindevkit.bdk'
 version = '1.0-SNAPSHOT'
 
 repositories {
