+++ uniffi.toml
@@ -1,0 +1,12 @@
+[bindings.kotlin]
+package_name = "org.bitcoindevkit"
+cdylib_name = "bdkffi"
+
+[bindings.python]
+cdylib_name = "bdkffi"
+
+[bindings.ruby]
+cdylib_name = "bdkffi"
+
+[bindings.swift]
+cdylib_name = "bdkffi"
